meilleur gestion des controllers et bon retour user sur le infinite scroll des amis
