+++ lib/component/form/login_form.dart
@@ -17,11 +17,8 @@ class LoginForm extends HookWidget {
 
   @override
   Widget build(BuildContext context) {
-    final TextEditingController emailController = useTextEditingController();
-    emailController.text = 'test@mail.com';
-    final TextEditingController passwordController = useTextEditingController();
-    passwordController.text = 'Test1234!';
-    useTextEditingController();
+    final TextEditingController emailController = useTextEditingController(text: 'test@mail.com');
+    final TextEditingController passwordController = useTextEditingController(text: 'Test1234!');
     final GlobalKey<FormState> formKey = GlobalKey<FormState>();
     return BlocProvider(
       create: (context) => LoginBloc(
